+++ tests/codegen_refs/max_pool2d.c
@@ -1,25 +1,26 @@
 #include <stdint.h>
+#include <stddef.h>
 #include <stdbool.h>
 #include "ops_scalar_f32.h"
 
 void node1_max_pool2d_f32(const float input[1][2][4][4], float out[1][2][3][3]) {
     int64_t out_h = 3;
     int64_t out_w = 3;
-    for (int64_t n = 0; n < 1; ++n) {
-        for (int64_t c = 0; c < 2; ++c) {
-            for (int64_t oh = 0; oh < out_h; ++oh) {
-                for (int64_t ow = 0; ow < out_w; ++ow) {
-                    int64_t in_h_base = oh * 1 - 0;
-                    int64_t in_w_base = ow * 1 - 0;
+    for (size_t n = 0; n < 1; ++n) {
+        for (size_t c = 0; c < 2; ++c) {
+            for (size_t oh = 0; oh < out_h; ++oh) {
+                for (size_t ow = 0; ow < out_w; ++ow) {
+                    int64_t in_h_base = (int64_t)oh * 1 - 0;
+                    int64_t in_w_base = (int64_t)ow * 1 - 0;
                     bool has_value = false;
                     float max_val = 0;
-                    for (int64_t kh = 0; kh < 2; ++kh) {
-                        int64_t in_h_idx = in_h_base + kh * 1;
+                    for (size_t kh = 0; kh < 2; ++kh) {
+                        int64_t in_h_idx = in_h_base + (int64_t)kh * 1;
                         if (in_h_idx < 0 || in_h_idx >= 4) {
                             continue;
                         }
-                        for (int64_t kw = 0; kw < 2; ++kw) {
-                            int64_t in_w_idx = in_w_base + kw * 1;
+                        for (size_t kw = 0; kw < 2; ++kw) {
+                            int64_t in_w_idx = in_w_base + (int64_t)kw * 1;
                             if (in_w_idx < 0 || in_w_idx >= 4) {
                                 continue;
                             }
