+++ tests/codegen_refs/mean_dim.c
@@ -1,12 +1,13 @@
 #include <stdint.h>
+#include <stddef.h>
 #include <stdbool.h>
 #include "ops_scalar_f32.h"
 
 void node1_mean_f32(const float a[2][3][4], float out[2][4]) {
-    for (int64_t i0 = 0; i0 < 2; ++i0) {
-        for (int64_t i1 = 0; i1 < 4; ++i1) {
+    for (size_t i0 = 0; i0 < 2; ++i0) {
+        for (size_t i1 = 0; i1 < 4; ++i1) {
             float acc = 0.0f;
-            for (int64_t r1 = 0; r1 < 3; ++r1) {
+            for (size_t r1 = 0; r1 < 3; ++r1) {
                 acc += a[i0][r1][i1];
             }
             acc /= (float)3;
