+++ tests/codegen_refs/sum_strided.c
@@ -1,11 +1,12 @@
 #include <stdint.h>
+#include <stddef.h>
 #include <stdbool.h>
 #include "ops_scalar_f32.h"
 
 void node1_sum_f32(const float a[4][3], float out[3]) {
-    for (int64_t i0 = 0; i0 < 3; ++i0) {
+    for (size_t i0 = 0; i0 < 3; ++i0) {
         float acc = 0.0f;
-        for (int64_t r0 = 0; r0 < 4; ++r0) {
+        for (size_t r0 = 0; r0 < 4; ++r0) {
             acc += ((float*)a)[r0 * 1 + i0 * 4];
         }
         out[i0] = acc;
